+++ src/ENDFtk/DirectoryRecord.hpp
@@ -50,7 +50,7 @@ public:
     return not( *this == rhs );
   }
 
-  int NC() const { return 1; }
+  long NC() const { return 1; }
   
   #include "ENDFtk/DirectoryRecord/src/print.hpp"
 };
