+++ src/ENDFtk/resonanceParameters/unresolved/EnergyIndependent.hpp
@@ -38,7 +38,7 @@ public:
         ( []( const auto& list ){ return Lvalue( list ); } );
   }
 
-  int NC() const {
+  long NC() const {
     return 2
       + ( ( this->ape ) ? this->ape->NC() : 0 )
       + ranges::accumulate
